more experiments

+++ experiments/branch_price_and_cut.jl
@@ -41,11 +41,11 @@ function run_experiment(out_dir, sizes, cuts, branching_rules, heuristic_time_li
         file_name = s * "_cut_" * string(cut) * "_branch_strategy_" * string(b_rule) * "_heuristic_" * string(heuristic_enabled)
         logfile = out_dir * "logs_" * file_name * ".txt"
 
-        # # check if we tried this file, skip if logs exist
-        # if (~precompile) && (isfile(logfile))
-        #     println("Skipping ", file_name)
-        #     continue
-        # end
+        # check if we tried this file, skip if logs exist
+        if (~precompile) && (isfile(logfile))
+            println("Skipping ", file_name)
+            continue
+        end
 
         local io = open(logfile, "w")
         if args["debug"] == true
@@ -93,16 +93,31 @@ branching_rules = ["most_fractional", "max_variance", "linking_dual_max_variance
 heuristic_time_limits = [0.0, 60.0]
 
 # precompile
-sizes = [(3, 10, 14, 26, 40.0 * 16.0)]
+sizes = [(3, 10, 14, 20, 640.0)]
 run_experiment(out_dir, sizes, cuts, branching_rules, heuristic_time_limits, precompile=true, total_time_limit=5.0)
 
-# experiment
+# sensitivity experiment
+branching_rules = ["linking_dual_max_variance"]
+heuristic_time_limits = [60.0]
 cuts = [true]
-# sizes = [(6, 20, 14, 20), (9, 30, 14, 20), (12, 40, 14, 20), (15, 50, 14, 20), (18, 60, 14, 20)]
 sizes_1 = [(6, 20, 14, i, j) for i ∈ [16, 18, 20, 22, 24] for j ∈ [640.0, 240.0, Inf]] 
 sizes_2 = [(9, 30, 14, i, j) for i ∈ [16, 18, 20, 22, 24] for j ∈ [640.0, 240.0, Inf]] 
 sizes = vcat(sizes_1, sizes_2)
-branching_rules = ["linking_dual_max_variance"]
-heuristic_time_limits = [60.0]
 run_experiment(out_dir, sizes, cuts, branching_rules, heuristic_time_limits, precompile=false, total_time_limit=1200.0)
 
+# value of cuts, branching, heuristic experiment
+cuts = [true, false]
+branching_rules = ["most_fractional", "max_variance", "linking_dual_max_variance"]
+heuristic_time_limits = [0.0, 60.0]
+sizes = [(3, 10, 14, 20, 640.0), (6, 20, 14, 20, 640.0)]
+run_experiment(out_dir, sizes, cuts, branching_rules, heuristic_time_limits, precompile=false, total_time_limit=1200.0)
+
+
+# scalability experiment
+branching_rules = ["linking_dual_max_variance"]
+heuristic_time_limits = [60.0]
+cuts = [true]
+branching_rules = ["most_fractional", "max_variance", "linking_dual_max_variance"]
+heuristic_time_limits = [0.0, 60.0]
+sizes = [(3, 10, 14, 20, 640.0), (6, 20, 14, 20, 640.0), (9, 30, 14, 20, 640.0), (12, 40, 14, 20, 640.0), (15, 50, 14, 20, 640.0), (18, 60, 14, 20, 640.0), (21, 70, 14, 20, 640.0)]
+run_experiment(out_dir, sizes, cuts, branching_rules, heuristic_time_limits, precompile=false, total_time_limit=1200.0)
